email_sending_cloud_function/requirements.txt
# Google Cloud Function dependencies
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-pubsub==2.*
google-auth==2.*
google-auth-oauthlib==1.*
google-auth-httplib2==0.*

# HTTP client and async support
aiohttp==3.*
aiolimiter>=1.0.0
requests>=2.28.0

# Data processing (if needed for analysis)
pandas>=1.5.0
numpy>=1.21.0

# Logging and utilities
structlog>=23.0.0
